Move board's validators to separate folder
