+++ src/Patronage.Api/Validators/Board/CreateBoardCommandValidator.cs
@@ -1,7 +1,7 @@
 ﻿using FluentValidation;
 using Patronage.Api.MediatR.Board.Commands.Create;
 
-namespace Patronage.Api.Validators
+namespace Patronage.Api.Validators.Board
 {
     public class CreateBoardCommandValidator : AbstractValidator<CreateBoardCommand>
     {
